Move Converter class to core directory
